Add dataset command now can import csvs with more than datasets. Separated the functionalities of dataframe and dataframeSchema.
