+++ mipdb/usecases.py
@@ -7,7 +7,8 @@ import pandas as pd
 
 from mipdb.database import DataBase, Connection
 from mipdb.database import METADATA_SCHEMA
-from mipdb.exceptions import ForeignKeyError
+from mipdb.data_frame_schema import DataFrameSchema
+from mipdb.exceptions import ForeignKeyError, InvalidDatasetError
 from mipdb.exceptions import UserInputError
 from mipdb.properties import Properties
 from mipdb.reader import CSVDataFrameReader
@@ -23,7 +24,7 @@ from mipdb.tables import (
     MetadataTable,
     PrimaryDataTable,
 )
-from mipdb.dataset import Dataset
+from mipdb.data_frame import DataFrame
 
 
 class UseCase(ABC):
@@ -178,10 +179,9 @@ class DeleteDataModel(UseCase):
         metadata = Schema(METADATA_SCHEMA)
         datasets_table = DatasetsTable(schema=metadata)
         with self.db.begin() as conn:
-            dataset_rows = datasets_table.get_datasets(
+            dataset_codes = datasets_table.get_datasets(
                 data_model_id=data_model_id, columns=["code"], db=conn
             )
-        dataset_codes = [dataset_row[0] for dataset_row in dataset_rows]
 
         for dataset_code in dataset_codes:
             DeleteDataset(self.db).execute(
@@ -208,62 +208,53 @@ class AddDataset(UseCase):
         with self.db.begin() as conn:
             metadata_table = MetadataTable.from_db(data_model, conn)
             dataset_enumerations = metadata_table.get_dataset_enums()
-            dataset_id = self._get_next_dataset_id(conn)
             data_model_id = data_model_table.get_data_model_id(
                 data_model_code, data_model_version, conn
             )
 
-            primary_data_table = PrimaryDataTable.from_db(data_model, conn)
-            with CSVDataFrameReader(csv_path).get_reader() as reader:
-                for dataset_data in reader:
-                    dataset = Dataset(dataset_data)
-
-                    dataset.validate_dataset(metadata_table.table)
-            self._verify_dataset_does_not_exist(
-                data_model_id=data_model_id,
-                dataset_name=dataset.name,
-                conn=conn,
+            ValidateDataset(self.db).execute(
+                csv_path, data_model_code, data_model_version
             )
-            with CSVDataFrameReader(csv_path).get_reader() as reader:
-                for dataset_data in reader:
-                    dataset = Dataset(dataset_data)
-                    values = dataset.to_dict()
-                    primary_data_table.insert_values(values, conn)
-            label = dataset_enumerations[dataset.name]
-
-            values = dict(
-                data_model_id=data_model_id,
-                dataset_id=dataset_id,
-                code=dataset.name,
-                label=label,
-                status="ENABLED",
+            imported_datasets = self._import_csv(
+                csv_path=csv_path, data_model=data_model, conn=conn
             )
-            datasets_table.insert_values(values, conn)
+            existing_datasets = datasets_table.get_datasets(columns=["code"], db=conn)
+            for dataset in imported_datasets:
+                if dataset not in existing_datasets:
+                    dataset_id = self._get_next_dataset_id(conn)
+                    values = dict(
+                        data_model_id=data_model_id,
+                        dataset_id=dataset_id,
+                        code=dataset,
+                        label=dataset_enumerations[dataset],
+                        status="ENABLED",
+                    )
+                    datasets_table.insert_values(values, conn)
+                    data_model_details = _get_data_model_details(data_model_id, conn)
+                    dataset_details = _get_dataset_details(dataset_id, conn)
+                    update_actions(
+                        conn=conn,
+                        action="ADD DATASET",
+                        data_model_details=data_model_details,
+                        dataset_details=dataset_details,
+                    )
 
-            data_model_details = _get_data_model_details(data_model_id, conn)
-            dataset_details = _get_dataset_details(dataset_id, conn)
-            update_actions(
-                conn=conn,
-                action="ADD DATASET",
-                data_model_details=data_model_details,
-                dataset_details=dataset_details,
-            )
+    def _import_csv(self, csv_path, data_model, conn):
+        imported_datasets = []
+        primary_data_table = PrimaryDataTable.from_db(data_model, conn)
+        with CSVDataFrameReader(csv_path).get_reader() as reader:
+            for dataset_data in reader:
+                dataframe = DataFrame(dataset_data)
+                imported_datasets = set(imported_datasets) | set(dataframe.datasets)
+                values = dataframe.to_dict()
+                primary_data_table.insert_values(values, conn)
+        return imported_datasets
 
     def _get_next_dataset_id(self, conn):
         metadata = Schema(METADATA_SCHEMA)
         datasets_table = DatasetsTable(schema=metadata)
         dataset_id = datasets_table.get_next_dataset_id(conn)
         return dataset_id
-
-    def _verify_dataset_does_not_exist(self, data_model_id, dataset_name, conn):
-        metadata = Schema(METADATA_SCHEMA)
-        dataset_table = DatasetsTable(schema=metadata)
-        datasets = dataset_table.get_datasets(
-            db=conn, data_model_id=data_model_id, columns=["code"]
-        )
-
-        if datasets and (dataset_name,) in datasets:
-            raise UserInputError("Dataset already exists!")
 
 
 class ValidateDataset(UseCase):
@@ -278,14 +269,30 @@ class ValidateDataset(UseCase):
         data_model = Schema(data_model_name)
 
         with self.db.begin() as conn:
-            self.validate_csv(csv_path, data_model, conn)
+            metadata_table = MetadataTable.from_db(data_model, conn)
+            dataset_enumerations = metadata_table.get_dataset_enums()
 
-    def validate_csv(self, csv_path, data_model, conn):
-        metadata_table = MetadataTable.from_db(data_model, conn)
-        with CSVDataFrameReader(csv_path).get_reader() as reader:
-            for dataset_data in reader:
-                dataset = Dataset(dataset_data)
-                dataset.validate_dataset(metadata_table.table)
+            dataframe_columns = pd.read_csv(csv_path, nrows=0).columns.tolist()
+            dataframe_schema = DataFrameSchema(
+                metadata_table=metadata_table.table, columns=dataframe_columns
+            )
+            with CSVDataFrameReader(csv_path).get_reader() as reader:
+                for dataset_data in reader:
+                    dataframe = DataFrame(dataset_data)
+                    dataframe_schema.validate_dataframe(dataframe.data)
+                    self.verify_datasets_exist_in_enumerations(
+                        datasets=dataframe.datasets,
+                        dataset_enumerations=dataset_enumerations,
+                    )
+
+    def verify_datasets_exist_in_enumerations(self, datasets, dataset_enumerations):
+        non_existing_datasets = [
+            dataset for dataset in datasets if dataset not in dataset_enumerations
+        ]
+        if non_existing_datasets:
+            raise InvalidDatasetError(
+                f"The values:'{non_existing_datasets}' are not present in the enumerations of the CDE 'dataset'."
+            )
 
 
 class DeleteDataset(UseCase):
