+++ tests/data/success/data_model_v_1_0/dataset123.csv
@@ -1,0 +1,6 @@
+subjectcode,var2,var3,var4,dataset
+2,1,11,,dataset
+2,2.0,12,22,dataset1
+3,1,13,23,dataset2
+4,1,14,24,dataset1
+3,2.0,15,25,dataset2
